// src/interface/interface.ts
interface ImageUrlInterface {
  link: string;
  name: string;
  isDefault?: boolean;
  type?: string;
}
interface SocialInterface {
  facebook?: string;
  twitter?: string;
  instagram?: string;
  linkedin?: string;
  youtube?: string;
  website?: string;
  tiktok?: string;
}

interface UserInterface {
  [key: string]: any;
  email: string;
  role: string;
  createdAt: Date;
  updatedAt: Date;
  _id: string;
  __v: number;
  firstName: string;
  lastName: string;
  displayName: string;
  phoneNumber: string;
  phoneNumberVerified: boolean;
  shopId: string;
  imageUrl: ImageUrlInterface;
  uid: string;
  userId: string;
  address: string;
  region: string;
  country: string;
  isVendor: boolean;
  disabled: boolean;
  isAdmin: boolean;
  superAdmin: boolean;
  social?: SocialInterface;
}
interface SighnUpInterface {
  email?: string;
  password?: string;
  firstName?: string;
  lastName?: string;
  phoneNumber?: string;
  address?: string;
  region?: string;
  country?: string;
  _id?: string;
  social?: SocialInterface;
}

interface CommentInterface {
  _id: string;
  userId?: string;
  comment: string;
  createdAt: Date;
  updatedAt: Date;
  __v: number;
  user: UserInterface;
  type: string;
  commentBy: UserInterface;
}